# backend/openapi.yaml
openapi: 3.0.0
info:
  version: 2.0.1
  title: Anki Clone Application API
  description: This is the API of the Anki Clone Application. It allows you to manage users, learning packages, learning facts, tags and statistics.
paths:
...
  /api/user:
    get:
      summary: Get all users (only for admin)
      operationId: getUser
      tags:
        - Users
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '500':
          description: Internal Server Error
    post:
      summary: Create a new user
      operationId: createUser
      tags:
        - Users
      requestBody:
        description: The user to create
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                  description: The mail of the user
                  example: john.doe@gmail.com
                  required: true
                pseudo:
                  type: string
                  description: The pseudo of the user
                  example: John
                  required: true
                password:
                  type: string
                  description: The password of the user
                  example: 123456
                  required: true
                profilePicture:
                  type: string
                  description: The profile picture of the user
                  example: https://www.google.com/
                  required: false

      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad Request
        '409':
          description: Conflict
        '500':
          description: Internal Server Error
  /api/user/{user_id}:
    get:
      summary: Get a user by id
      operationId: getUserById
      tags:
        - Users
      parameters:
        - name: user_id
          in: path
          description: The id of the user to retrieve
          required: true
          schema:
            type: number
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
    put:
      summary: Update a user by id
      operationId: updateUserById
      tags:
        - Users
      parameters:
        - name: user_id
          in: path
          description: The id of the user to update
          required: true
          schema:
            type: number
      requestBody:
        description: The user to update
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                  description: The mail of the user
                  example: john.doe@gmail.com
                pseudo:
                  type: string
                  description: The pseudo of the user
                  example: John
                password:
                  type: string
                  description: The password of the user
                  example: 123456
                profilePicture:
                  type: string
                  description: The profile picture of the user
                  example: https://www.google.com/
      responses:
        '204':
          description: Updated
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /api/user/{user_id}/deactivate:
    put:
      summary: Deactivate a user by id
      operationId: deactivateUserById
      tags:
        - Users
      parameters:
        - name: user_id
          in: path
          description: The id of the user to deactivate
          required: true
          schema:
            type: number
      responses:
        '204':
          description: Updated
        '404':
          description: Not Found
        '500':
          description: Internal Server Error

  # LEARNING PACKAGE API
...
  /api/statistics/user/{user_id}:
    get:
      summary: Get the statistics of a user by id
      operationId: getStatisticsByUserId
      tags:
        - Statistics
      parameters:
        - name: user_id
          in: path
          description: The id of the user
          required: true
          schema:
            type: number
      responses:
        '200':
          description: A list of user statistics
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    nextDays:
                      type: object
                      additionalProperties:
                        type: integer
                    todayFacts:
                      type: integer
                    globalProgress:
                      type: number
                      format: double
                    nbCreatedPackages:
                      type: integer
        '404':
          description: Not Found
        '500':
          description: Internal Server Error

  # LEARNING PACKAGE <> FACT API
...
  /api/package/user/{user_id}:
    get:
      summary: Get all learning packages by user id
      operationId: getLearningPackageByUserId
      tags:
        - Learning Packages <> Users
      parameters:
        - name: user_id
          in: path
          description: The id of the user
          required: true
          schema:
            type: number
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    packageId:
                      type: number
                      description: The id of the learning package
                      example: 1
                    title:
                      type: string
                      description: The title of the learning package
                      example: Learn TypeScript
                    description:
                      type: string
                      description: The description of the learning package
                      example: A comprehensive guide to understanding TypeScript, a statically typed superset of JavaScript.
                    category:
                      type: string
                      description: The category of the learning package
                      example: Programming Language
                    targetAudience:
                      type: string
                      description: The target audience of the learning package
                      example: Developers familiar with JavaScript
                    duration:
                      type: number
                      description: The duration of the learning package in hours
                      example: 2
                    creationDate:
                      type: string
                      description: The creation date of the learning package
                      example: 2021-01-01
                    creatorId:
                      type: number
                      description: The id of the creator of the learning package
                      example: 1
                    progress:
                      type: number
                      description: The progress of the user learning package
                      example: 0
                    expectedEndDate:
                      type: string
                      description: The expected end date of the user learning package
                      example: 2021-01-01
                    minutesPerDayObjective:
                      type: number
                      description: The minutes per day objective of the user learning package
                      example: 30
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
...
  /api/fact/user/{user_id}:
    get:
      summary: Get all learning facts by user id
      operationId: getLearningFactByUserId
      tags:
        - Learning Facts <> Users
      parameters:
        - name: user_id
          in: path
          description: The id of the user
          required: true
          schema:
            type: number
      responses:
        '200':
          description: A list of learning facts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserLearningFact'
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /api/fact/{fact_id}/user/{user_id}:
    get:
      summary: Get a learning fact by id and user id
      operationId: getLearningFactByIdAndUserId
      tags:
        - Learning Facts <> Users
      parameters:
        - name: fact_id
          in: path
          description: The id of the learning fact
          required: true
          schema:
            type: number
        - name: user_id
          in: path
          description: The id of the user
          required: true
          schema:
            type: number
      responses:
        '200':
          description: A learning fact
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserLearningFact'
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
...
servers:
  [
    {
      url: http://localhost:3000,
      description: Development server
    }
  ]
components:
  schemas:
    User:
      type: object
      properties:
        userId:
          type: number
          description: The id of the user
          example: 1
        mail:
          type: string
          description: The mail of the user
          example: john.doe@gmail.com
        pseudo:
          type: string
          description: The pseudo of the user
          example: John
        password:
          type: string
          description: The password of the user
          example: 123456
        registrationDate:
          type: string
          description: The registration date of the user
          example: 2021-01-01
        profilePicture:
          type: string
          description: The profile picture of the user
          example: https://www.google.com/
        isActive:
          type: boolean
          description: The status of the user
          example: true
      required:
        - userId
        - mail
        - pseudo
        - password
        - registrationDate
        - isActive
...
    UserLearningFact:
      type: object
      properties:
        userLearningFactId:
          type: number
          description: The id of the user learning fact
          example: 1
        reviewCount:
          type: number
          description: The review count of the user learning fact
          example: 1
        confidenceLevel:
          type: string
          description: The confidence level of the user learning fact
          example: 1
        lastReviewed:
          type: string
          description: The last reviewed date of the user learning fact
          example: 2021-01-01
        nextReviewDate:
          type: string
          description: The next review date of the user learning fact
          example: 2021-01-01
        factId:
          type: number
          description: The id of the learning fact
          example: 1
        userId:
          type: number
          description: The id of the user
          example: 1
      required:
        - userLearningFactId
        - reviewCount
        - confidenceLevel
        - lastReviewed
        - nextReviewDate
        - factId
        - userId
...